delete auth for create user and item and hide edit and delete btn for non-admin user

+++ src/resolvers/Mutation.js
@@ -1,4 +1,4 @@
-import moment from 'moment'
+import moment from "moment";
 import bcrypt from "bcryptjs";
 import hashPassword from "../utils/hashPassword";
 import generateToken from "../utils/generateToken";
@@ -6,9 +6,6 @@ import getUserId from "../utils/getUserId";
 
 const Mutation = {
   async createUser(parent, args, { prisma, request }, info) {
-    if (args.auth !== "Eg80949597") {
-      throw new Error("Authentication required.");
-    }
     const employeeId = getUserId(request);
     const user = await prisma.query.user(
       {
@@ -38,7 +35,7 @@ const Mutation = {
     }
   },
   async login(parent, args, { prisma }, info) {
-    const { employeeId, password } = args.data
+    const { employeeId, password } = args.data;
     const oldUser = await prisma.query.user({
       where: {
         employeeId,
@@ -57,13 +54,13 @@ const Mutation = {
 
     const user = await prisma.mutation.updateUser({
       where: {
-        employeeId
+        employeeId,
       },
       data: {
         loginTime: moment().format().toString(),
-        lastLogin: oldUser.loginTime
-      }
-    })
+        lastLogin: oldUser.loginTime,
+      },
+    });
     return {
       user,
       token: generateToken(user.employeeId),
@@ -205,16 +202,19 @@ const Mutation = {
   async deleteFreetimes(parent, args, { prisma, request }, info) {
     const employeeId = getUserId(request);
     if (employeeId) {
-      return prisma.mutation.updateManyFreetimes({
-        where: {
-          schedule: {
-            schedule_No: args.schedule_No
-          }
+      return prisma.mutation.updateManyFreetimes(
+        {
+          where: {
+            schedule: {
+              schedule_No: args.schedule_No,
+            },
+          },
+          data: {
+            availability: "full",
+          },
         },
-        data: {
-          availability: "full"
-        }
-      }, info);
+        info
+      );
     } else {
       throw new Error("Permission denied.");
     }
@@ -224,7 +224,7 @@ const Mutation = {
 
   async createSchedule(parent, args, { prisma, request }, info) {
     let count = 0;
-    const { schedule_No } = args
+    const { schedule_No } = args;
     //create schedule
     await prisma.mutation.createSchedule({
       data: {
@@ -379,21 +379,21 @@ const Mutation = {
     return { count: 0 };
   },
   deleteSchedule_Staffs(parent, args, { prisma, request }, info) {
-    return prisma.mutation.deleteManySchedule_Staffs({
-      where: {
-        schedule: {
-          schedule_No: args.schedule_No
-        }
-      }
-    }, info)
+    return prisma.mutation.deleteManySchedule_Staffs(
+      {
+        where: {
+          schedule: {
+            schedule_No: args.schedule_No,
+          },
+        },
+      },
+      info
+    );
   },
 
   //---------------------------  Item ----------------------------//
 
   async createItem(parent, args, { prisma, request }, info) {
-    if (args.auth !== "Eg80949597") {
-      throw new Error("Authentication required.");
-    }
     const employeeId = getUserId(request);
     const user = await prisma.query.user(
       {
@@ -405,13 +405,13 @@ const Mutation = {
     );
     if (user.accountType === "Admin") {
       const isExist = await prisma.exists.Item({
-        description_en: args.data.description_en
-      })
+        description_en: args.data.description_en,
+      });
       if (isExist) {
         throw new Error("description taken!");
       }
       return prisma.mutation.createItem({
-        data: args.data
+        data: args.data,
       });
     } else {
       throw new Error("Permission denied.");
@@ -441,15 +441,28 @@ const Mutation = {
       throw new Error("Permission denied.");
     }
   },
-  deleteItem(parent, args, { prisma, request }, info) {
-    return prisma.mutation.deleteItem(
+  async deleteItem(parent, args, { prisma, request }, info) {
+    const employeeId = getUserId(request);
+    const user = await prisma.query.user(
       {
         where: {
-          id: args.id,
+          employeeId,
         },
       },
-      info
+      "{ accountType }"
     );
+    if (user.accountType === "Admin") {
+      return prisma.mutation.deleteItem(
+        {
+          where: {
+            id: args.id,
+          },
+        },
+        info
+      );
+    } else {
+      throw new Error("Permission denied.");
+    }
   },
 };
